chore: Update npm dependencies and add Sass for styling
