+++ package.json
@@ -9,6 +9,7 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-scripts": "5.0.1",
+    "sass": "^1.77.2",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
