+++ src/components/searchBar/SearchBar.module.scss
@@ -1,0 +1,55 @@
+@import "../../sass/common";
+
+.searchBar {
+    border: none;
+    height: 3rem;
+    background-color: map-get($color-theme, "light");
+    padding: 1rem;
+    border-radius: 10px;
+    min-width: 15rem;
+    width: 40rem;
+    max-width: 100%;
+    margin-right: 1rem;
+    height: 4rem;
+    color: map-get($color-theme, "primary");
+    box-shadow: inset 5px 5px 5px 0 rgba(map-get($color-theme, "dark"), 0.2),
+        inset -3px -3px 5px 0 rgba(255, 255, 255, 0.5);
+
+    &:focus {
+        outline: none;
+    }
+    &::placeholder {
+        color: map-get($color-theme, "primary");
+    }
+}
+
+.button {
+    padding: 1rem;
+    border-radius: 10px;
+    border: none;
+    height: 4rem;
+    background-color: map-get($color-theme, "light");
+    font-weight: 500;
+    color: map-get($color-theme, "primary");
+    box-shadow: 5px 5px 2px rgba(map-get($color-theme, "dark"), 0.1),
+        inset 5px 5px 5px 0 rgba(255, 255, 255, 0.2),
+        inset -3px -3px 5px 0 rgba(map-get($color-theme, "dark"), 0.5);
+    transition: all 0.3s ease;
+
+    &:hover {
+        background-color: map-get($color-theme, "secondary");
+        color: map-get($color-theme, "light");
+        transform: translateY(-0.2rem);
+        box-shadow: 7px 7px 5px rgba(map-get($color-theme, "dark"), 0.1),
+            inset 5px 5px 5px 0 rgba(255, 255, 255, 0.2),
+            inset -3px -3px 5px 0 rgba(map-get($color-theme, "dark"), 0.5);
+    }
+
+    &:active {
+        background-color: map-get($color-theme, "tertiary");
+        color: map-get($color-theme, "light");
+        transform: translateY(0.2rem);
+        box-shadow: inset 5px 5px 5px 0 rgba(map-get($color-theme, "dark"), 0.5),
+            inset -3px -3px 5px 0 rgba(map-get($color-theme, "dark"), 0.2);
+    }
+}
