--- src/pages/Home.jsx
@@ -1,13 +1,0 @@
-import React from "react";
-
-import SearchBarContainer from '../containers/SearchBarContainer'
-
-const Home = () => {
-    return (
-        <div>
-            <SearchBarContainer />
-        </div>
-    )
-}
-
-export default Home;
