+++ src/pages/home/Home.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+import Header from "../../components/header";
+
+const Home = () => {
+    return (
+        <>
+            <Header />
+        </>
+    );
+};
+
+export default Home;
