+++ src/sass/_common.scss
@@ -1,0 +1,8 @@
+$color-theme: (
+    primary: rgba(3, 172, 191, 1),
+    secondary: rgba(3, 191, 191, 1),
+    tertiary: rgba(1, 89, 89, 1),
+    accent: rgba(160, 165, 3, 1),
+    light: rgb(238, 233, 227),
+    dark: rgb(1, 0, 38),
+);
